+++ shardtree/src/lib.rs
@@ -641,7 +641,7 @@ impl<
     /// given position, and parents of such nodes, will be replaced by the empty root for the
     /// associated level.
     ///
-    /// Use [`Self::root_at_checkpoint`] to obtain the root of the overall tree.
+    /// Use [`Self::root_at_checkpoint_depth`] to obtain the root of the overall tree.
     pub fn root(
         &self,
         address: Address,
@@ -970,12 +970,47 @@ impl<
         })
     }
 
+    /// Computes the root of the tree as of the checkpointed position having the specified
+    /// checkpoint id.
+    pub fn root_at_checkpoint_id(&self, checkpoint: &C) -> Result<H, ShardTreeError<S::Error>> {
+        let position = self
+            .store
+            .get_checkpoint(checkpoint)
+            .map_err(ShardTreeError::Storage)?
+            .map(|c| c.position())
+            .ok_or(QueryError::CheckpointPruned)?;
+
+        position.map_or_else(
+            || Ok(H::empty_root(Self::root_addr().level())),
+            |pos| self.root(Self::root_addr(), pos + 1),
+        )
+    }
+
+    /// Computes the root of the tree as of the checkpointed position having the specified
+    /// checkpoint id, caching intermediate values produced while computing the root.
+    pub fn root_at_checkpoint_id_caching(
+        &mut self,
+        checkpoint: &C,
+    ) -> Result<H, ShardTreeError<S::Error>> {
+        let position = self
+            .store
+            .get_checkpoint(checkpoint)
+            .map_err(ShardTreeError::Storage)?
+            .map(|c| c.position())
+            .ok_or(QueryError::CheckpointPruned)?;
+
+        position.map_or_else(
+            || Ok(H::empty_root(Self::root_addr().level())),
+            |pos| self.root_caching(Self::root_addr(), pos + 1),
+        )
+    }
+
     /// Computes the root of the tree as of the checkpointed position at the specified depth.
     ///
     /// Returns the root as of the most recently appended leaf if `checkpoint_depth == 0`. Note
     /// that if the most recently appended leaf is also a checkpoint, this will return the same
     /// result as `checkpoint_depth == 1`.
-    pub fn root_at_checkpoint(
+    pub fn root_at_checkpoint_depth(
         &self,
         checkpoint_depth: usize,
     ) -> Result<H, ShardTreeError<S::Error>> {
@@ -985,7 +1020,9 @@ impl<
         )
     }
 
-    pub fn root_at_checkpoint_caching(
+    /// Computes the root of the tree as of the checkpointed position at the specified depth,
+    /// caching intermediate values produced while computing the root.
+    pub fn root_at_checkpoint_depth_caching(
         &mut self,
         checkpoint_depth: usize,
     ) -> Result<H, ShardTreeError<S::Error>> {
@@ -995,92 +1032,146 @@ impl<
         )
     }
 
-    /// Computes the witness for the leaf at the specified position.
+    fn witness_helper<Ctx>(
+        mut ctx: Ctx,
+        position: Position,
+        as_of: Position,
+        get_shard: impl Fn(&Ctx, Address) -> Result<Option<LocatedPrunableTree<H>>, S::Error>,
+        mut root: impl FnMut(&mut Ctx, Address, Position) -> Result<H, ShardTreeError<S::Error>>,
+    ) -> Result<MerklePath<H, DEPTH>, ShardTreeError<S::Error>> {
+        let subtree_addr = Self::subtree_addr(position);
+
+        // compute the witness for the specified position up to the subtree root
+        let mut witness = get_shard(&ctx, subtree_addr)
+            .map_err(ShardTreeError::Storage)?
+            .map_or_else(
+                || Err(QueryError::TreeIncomplete(vec![subtree_addr])),
+                |subtree| subtree.witness(position, as_of + 1),
+            )?;
+
+        // compute the remaining parts of the witness up to the root
+        let root_addr = Self::root_addr();
+        let mut cur_addr = subtree_addr;
+        while cur_addr != root_addr {
+            witness.push(root(&mut ctx, cur_addr.sibling(), as_of + 1)?);
+            cur_addr = cur_addr.parent();
+        }
+
+        Ok(MerklePath::from_parts(witness, position).unwrap())
+    }
+
+    fn witness_internal(
+        &self,
+        position: Position,
+        as_of: Position,
+    ) -> Result<MerklePath<H, DEPTH>, ShardTreeError<S::Error>> {
+        Self::witness_helper(
+            self,
+            position,
+            as_of,
+            |ctx, shard_root| ctx.store.get_shard(shard_root),
+            |ctx, address, truncate_at| ctx.root(address, truncate_at),
+        )
+    }
+
+    fn witness_internal_caching(
+        &mut self,
+        position: Position,
+        as_of: Position,
+    ) -> Result<MerklePath<H, DEPTH>, ShardTreeError<S::Error>> {
+        Self::witness_helper(
+            self,
+            position,
+            as_of,
+            |ctx, shard_root| ctx.store.get_shard(shard_root),
+            |ctx, address, truncate_at| ctx.root_caching(address, truncate_at),
+        )
+    }
+
+    /// Computes the witness for the leaf at the specified position, as of the given checkpoint
+    /// depth.
     ///
     /// Returns the witness as of the most recently appended leaf if `checkpoint_depth == 0`. Note
     /// that if the most recently appended leaf is also a checkpoint, this will return the same
     /// result as `checkpoint_depth == 1`.
-    pub fn witness(
+    pub fn witness_at_checkpoint_depth(
         &self,
         position: Position,
         checkpoint_depth: usize,
     ) -> Result<MerklePath<H, DEPTH>, ShardTreeError<S::Error>> {
-        let max_leaf_position = self.max_leaf_position(checkpoint_depth).and_then(|v| {
+        let as_of = self.max_leaf_position(checkpoint_depth).and_then(|v| {
             v.ok_or_else(|| QueryError::TreeIncomplete(vec![Self::root_addr()]).into())
         })?;
 
-        if position > max_leaf_position {
+        if position > as_of {
             Err(
                 QueryError::NotContained(Address::from_parts(Level::from(0), position.into()))
                     .into(),
             )
         } else {
-            let subtree_addr = Self::subtree_addr(position);
-
-            // compute the witness for the specified position up to the subtree root
-            let mut witness = self
-                .store
-                .get_shard(subtree_addr)
-                .map_err(ShardTreeError::Storage)?
-                .map_or_else(
-                    || Err(QueryError::TreeIncomplete(vec![subtree_addr])),
-                    |subtree| subtree.witness(position, max_leaf_position + 1),
-                )?;
-
-            // compute the remaining parts of the witness up to the root
-            let root_addr = Self::root_addr();
-            let mut cur_addr = subtree_addr;
-            while cur_addr != root_addr {
-                witness.push(self.root(cur_addr.sibling(), max_leaf_position + 1)?);
-                cur_addr = cur_addr.parent();
-            }
-
-            Ok(MerklePath::from_parts(witness, position).unwrap())
+            self.witness_internal(position, as_of)
         }
     }
 
-    /// Computes the witness for the leaf at the specified position.
+    /// Computes the witness for the leaf at the specified position, as of the given checkpoint
+    /// depth.
     ///
     /// This implementation will mutate the tree to cache intermediate root (ommer) values that are
     /// computed in the process of constructing the witness, so as to avoid the need to recompute
     /// those values from potentially large numbers of subtree roots in the future.
-    pub fn witness_caching(
+    pub fn witness_at_checkpoint_depth_caching(
         &mut self,
         position: Position,
         checkpoint_depth: usize,
     ) -> Result<MerklePath<H, DEPTH>, ShardTreeError<S::Error>> {
-        let max_leaf_position = self.max_leaf_position(checkpoint_depth).and_then(|v| {
+        let as_of = self.max_leaf_position(checkpoint_depth).and_then(|v| {
             v.ok_or_else(|| QueryError::TreeIncomplete(vec![Self::root_addr()]).into())
         })?;
 
-        if position > max_leaf_position {
+        if position > as_of {
             Err(
                 QueryError::NotContained(Address::from_parts(Level::from(0), position.into()))
                     .into(),
             )
         } else {
-            let subtree_addr = Address::above_position(Self::subtree_level(), position);
-
-            // compute the witness for the specified position up to the subtree root
-            let mut witness = self
-                .store
-                .get_shard(subtree_addr)
-                .map_err(ShardTreeError::Storage)?
-                .map_or_else(
-                    || Err(QueryError::TreeIncomplete(vec![subtree_addr])),
-                    |subtree| subtree.witness(position, max_leaf_position + 1),
-                )?;
-
-            // compute the remaining parts of the witness up to the root
-            let root_addr = Self::root_addr();
-            let mut cur_addr = subtree_addr;
-            while cur_addr != root_addr {
-                witness.push(self.root_caching(cur_addr.sibling(), max_leaf_position + 1)?);
-                cur_addr = cur_addr.parent();
-            }
-
-            Ok(MerklePath::from_parts(witness, position).unwrap())
+            self.witness_internal_caching(position, as_of)
         }
+    }
+
+    /// Computes the witness for the leaf at the specified position, as of the given checkpoint.
+    pub fn witness_at_checkpoint_id(
+        &self,
+        position: Position,
+        checkpoint_id: &C,
+    ) -> Result<MerklePath<H, DEPTH>, ShardTreeError<S::Error>> {
+        let as_of = self
+            .store
+            .get_checkpoint(checkpoint_id)
+            .map_err(ShardTreeError::Storage)?
+            .and_then(|c| c.position())
+            .ok_or(QueryError::CheckpointPruned)?;
+
+        self.witness_internal(position, as_of)
+    }
+
+    /// Computes the witness for the leaf at the specified position, as of the given checkpoint.
+    ///
+    /// This implementation will mutate the tree to cache intermediate root (ommer) values that are
+    /// computed in the process of constructing the witness, so as to avoid the need to recompute
+    /// those values from potentially large numbers of subtree roots in the future.
+    pub fn witness_at_checkpoint_id_caching(
+        &mut self,
+        position: Position,
+        checkpoint_id: &C,
+    ) -> Result<MerklePath<H, DEPTH>, ShardTreeError<S::Error>> {
+        let as_of = self
+            .store
+            .get_checkpoint(checkpoint_id)
+            .map_err(ShardTreeError::Storage)?
+            .and_then(|c| c.position())
+            .ok_or(QueryError::CheckpointPruned)?;
+
+        self.witness_internal_caching(position, as_of)
     }
 
     /// Make a marked leaf at a position eligible to be pruned.
@@ -1341,8 +1432,8 @@ mod tests {
                 assert_eq!(root, caching_root);
 
                 for pos in marked_positions {
-                    let witness = tree.witness(pos, 0);
-                    let caching_witness = tree.witness_caching(pos, 0);
+                    let witness = tree.witness_at_checkpoint_depth(pos, 0);
+                    let caching_witness = tree.witness_at_checkpoint_depth_caching(pos, 0);
                     assert_matches!(witness, Ok(_));
                     assert_eq!(witness, caching_witness);
                 }